Move pog library classes to separate namespace
