+++ src/co/tkjn/gametable/poglibrary/PogFilenameFilter.java
@@ -3,7 +3,7 @@
  */
 
 
-package co.tkjn.gametable;
+package co.tkjn.gametable.poglibrary;
 
 import java.io.File;
 import java.io.FilenameFilter;
